fix: (tests) Fix tests
